peer comments resolved. Merged and tidied up Molecules/Functions folder. Will begin working on Pages/HomePage
