+++ src/Components/Pages/HomePage/index.tsx
@@ -1,9 +1,9 @@
+import UserHeaderMolecule from "../../Molecules/UserHeaderMolecule";
+
+
 function HomePage() {
     return (
-        <>
-            <h1>HomePage</h1>
-            <p>Will show all user information</p>
-    </>
+            <UserHeaderMolecule />
 )
 }
